requirements.txt
streamlit
langchain
langchain_community
python-dotenv
faiss-gpu